Refactor to facilitate local development of consistent rate code --- run with::
python src/vivarium_nih_moud/data/dismod_at.py
